requirements.txt
flask==2.0.0
python-dotenv==1.0.0
redis==4.5.4